Removing deprecated views, clean up and re-organising structure of the project
