+++ SampleApps/PingExample/PingExample/ViewModels/ConfigurationViewModel.swift
