+++ SampleApps/PingExample/PingExample/ViewModels/LoginViewModel.swift
