+++ SampleApps/PingExample/PingExample/ViewModels/StorageViewModel.swift
