+++ SampleApps/PingExample/PingExample/Views/InputView.swift
