+++ SampleApps/PingExample/PingExample/Views/StorageView.swift
@@ -1,0 +1,27 @@
+// 
+//  StorageView.swift
+//  PingExample
+//
+//  Copyright (c) 2024 Ping Identity. All rights reserved.
+//
+//  This software may be modified and distributed under the terms
+//  of the MIT license. See the LICENSE file for details.
+//
+
+import SwiftUI
+
+struct StorageView: View {
+    var storageViewModel = StorageViewModel()
+    var body: some View {
+        Text("This View is for testing Storage functionality.\nPlease check the Console Logs")
+            .font(.title3)
+            .multilineTextAlignment(.center)
+            .navigationBarTitle("Storage", displayMode: .inline)
+            .onAppear() {
+                Task {
+                    await storageViewModel.setupMemoryStorage()
+                    await storageViewModel.setupKeychainStorage()
+                }
+            }
+    }
+}
